engine: Fixes data race in barrier
deprecate WaitGroupBarrier:
rules of wg.Add: (https://golang.org/pkg/sync/#WaitGroup)

Note that calls with a positive delta that occur when the counter is zero must happen before a Wait. Calls with a negative delta, or calls with a positive delta that start when the counter is greater than zero, may happen at any time. Typically this means the calls to Add should execute before the statement creating the goroutine or other event to be waited for. If a WaitGroup is reused to wait for several independent sets of events, new Add calls must happen after all previous Wait calls have returned. See the WaitGroup example.

==> so for barrier, if Wait() happens before Pause(), it breaks the rule. However in read-world, it is not guareented socket upgrading happens before any other i/o;

the lockBarrier is more safe